move slugger out to a different file
